v4/iamwatching/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iamwatching"
version = "2.0.0"
description = "Multi-cloud IAM security auditor: shadow admin detection, cross-cloud credential discovery, Neo4j graph analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Aniza Corp Security Research" }]
keywords = ["iam", "security", "aws", "azure", "gcp", "neo4j", "cspm", "cloud-security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
]

dependencies = [
    # ── CLI / Output ────────────────────────────────────────────────────────
    "click>=8.1",
    "rich>=13.7",
    "pyyaml>=6.0",
    # ── Async HTTP ──────────────────────────────────────────────────────────
    "aiohttp>=3.9",
    "reportlab>=4.0",
    # ── Neo4j graph backend ──────────────────────────────────────────────────
    "neo4j>=5.20",
    # ── AWS scanning ────────────────────────────────────────────────────────
    "aioboto3>=13.1",
    "aiobotocore>=2.13",
    "botocore>=1.34",
    # ── Azure scanning ──────────────────────────────────────────────────────
    "azure-identity>=1.17",
    "azure-mgmt-authorization>=4.0",
    "azure-mgmt-compute>=32.0",
    "azure-mgmt-resource>=23.0",
    "azure-mgmt-web>=7.3",
    "msgraph-sdk>=1.0",
    # ── GCP scanning ────────────────────────────────────────────────────────
    "google-auth>=2.30",
    "google-auth-httplib2>=0.2",
    "google-cloud-functions>=1.16",
    "google-cloud-run>=0.10.9",
    "google-cloud-secret-manager>=2.20",
    "google-api-python-client>=2.130",
    "google-cloud-compute>=1.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.3",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "moto[iam,s3,lambda,ec2,sts]>=5.0",
    "black>=24.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pyinstaller>=6.8",
]
all = ["iamwatching[dev]"]

[project.scripts]
iamwatching = "iamwatching.cli.main:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["iamwatching*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths    = ["tests"]

[tool.ruff]
line-length    = 100
target-version = "py311"

[tool.mypy]
python_version      = "3.11"
strict              = false
ignore_missing_imports = true
